Time: 31 ms (62.54%), Space: 61.5 MB (16.43%) - LeetHub

+++ 2356-largest-combination-with-bitwise-and-greater-than-zero/2356-largest-combination-with-bitwise-and-greater-than-zero.cpp
@@ -60,7 +60,7 @@ public:
         for (int& val : candidates)
         {
             bitset<24> bits (val);
-            for (int i = 0; i < 24; i++)
+            for (short i = 0; i < 24; i++)
             {
                 if(bits[i])
                 ones[i]++;
